Extract shared API contract and fix Docker builds
