+++ .dockerignore
@@ -7,6 +7,7 @@
 !pnpm-workspace.yaml
 !.npmrc
 !packages/tsconfig/
+!packages/api-contract/
 !packages/dxdata/
 !packages/maimai-domain/
 !apps/backend/
